+++ src/misc/zx81/zx81basicvars.ts
@@ -18,11 +18,11 @@ export class Zx81BasicVars {
 		this.basicVars.clear();
 
 		// Parse the BASIC vars
-		let i = 0;
-		while (i < basicVars.length) {
+		let i = 0; const len = basicVars.length;
+		while (i < len) {
 			// Check for type
 			const firstByte = basicVars[i++];
-			let firstLetter = firstByte & 0b0001_1111;
+			let firstLetter = (firstByte & 0b0001_1111) + 0x20;
 			const type = firstByte & 0b1110_0000;
 			switch (type) {
 				case 0b0110_0000:	// One letter number
@@ -75,28 +75,28 @@ export class Zx81BasicVars {
 					}
 					break;
 
-				case 0b0100_0000:	// Array of numbers (one letter only)
+				case 0b1000_0000:	// Array of numbers (one letter only)
 					{
 						// Letter
-						firstLetter += 0x20;
 						if (firstLetter < 0x26 || firstLetter > 0x3F)
 							throw Error('Invalid BASIC variable.');
 						const varName = Zx81Tokens.tokens[firstLetter];
 						// Read array total length
 						const arrayLength = basicVars[i++] + 256 * basicVars[i++];
 						// Number of dimensions
-						const dimensions = basicVars[i++];	// +1 ?? TODO check
+						const dimensions = basicVars[i++];
+						const floatsLength = arrayLength - 1 - 2 * dimensions;
 						// Read size of each dimension
 						const dimSizes: number[] = [];
 						const dimIndex: number[] = [];
 						for (let j = 0; j < dimensions; j++) {
 							const dimSize = basicVars[i++] + 256 * basicVars[i++];
 							dimSizes.push(dimSize);
-							dimIndex.push(0);
+							dimIndex.push(1);
 						}
 						// Read values
 						let k = 0;
-						while (k < arrayLength) {
+						while (k < floatsLength) {
 							const valueBuf = basicVars.slice(i, i + 5);
 							const value = Zx81Tokens.convertBufToZx81Float(valueBuf);
 							// Store the variable
@@ -106,9 +106,9 @@ export class Zx81BasicVars {
 							let j = dimensions - 1;
 							while (j >= 0) {
 								dimIndex[j]++;
-								if (dimIndex[j] < dimSizes[j])
+								if (dimIndex[j] <= dimSizes[j])
 									break;
-								dimIndex[j] = 0;
+								dimIndex[j] = 1;
 								j--;
 							}
 							i += 5;
@@ -120,7 +120,6 @@ export class Zx81BasicVars {
 				case 0b0100_0000:	// String (one letter only)
 					{
 						// Letter, number of chars, chars
-						firstLetter += 0x20;
 						if (firstLetter < 0x26 || firstLetter > 0x3F)
 							throw Error('Invalid BASIC variable.');
 						const varName = Zx81Tokens.tokens[firstLetter] + '$';
@@ -140,7 +139,6 @@ export class Zx81BasicVars {
 				case 0b1100_0000:	// Array of chars (one letter only)
 					{
 						// Letter
-						firstLetter += 0x20;
 						if (firstLetter < 0x26 || firstLetter > 0x3F)
 							throw Error('Invalid BASIC variable.');
 						const varName = Zx81Tokens.tokens[firstLetter];
@@ -194,5 +192,17 @@ export class Zx81BasicVars {
 		}
 		const result = results.join(', ');
 		return result;
+	}
+
+
+	/** Returns all variables with values.
+	 * @returns A string with all variables and values.
+	 */
+	public getAllVariablesWithValues(): string[] {
+		const results: string[] = [];
+		for (let [varName, value] of this.basicVars) {
+			results.push(varName + '=' + (value === undefined ? 'undefined' : value));
+		}
+		return results;
 	}
 }